rename collectionsIndex to collectionsList
