+++ src/navigation/RootNavigator.js
@@ -9,7 +9,7 @@ import { shape } from 'prop-types';
 import LoginScreen from '../screens/LoginScreen';
 import SummaryScreen from '../screens/SummaryScreen';
 import CatalogScreen from '../screens/CatalogScreen';
-import CollectionsIndexScreen from '../screens/CollectionsIndexScreen';
+import CollectionsListScreen from '../screens/CollectionsListScreen';
 import CollectionPreviewScreen from '../screens/CollectionPreviewScreen';
 import Colors from '../constants/Colors';
 
@@ -43,12 +43,12 @@ const selectScreenOptions = ({ route: { name } }) => ({
 
 const CollectionStackScreen = () => (
   <CollectionsStack.Navigator
-    initialRouteName="CollectionsIndex"
+    initialRouteName="CollectionsList"
     headerMode="none"
   >
     <CollectionsStack.Screen
-      name="CollectionsIndex"
-      component={CollectionsIndexScreen}
+      name="CollectionsList"
+      component={CollectionsListScreen}
     />
     <CollectionsStack.Screen
       name="CollectionPreview"
